Revert "[Fleet] Fix get file handler for bundled package (#172182)"
This reverts commit 4d52ad2cd5d269c296af12a1fbf549ba8c1db95a.
