+++ x-pack/plugins/fleet/tsconfig.json
@@ -102,6 +102,5 @@
     "@kbn/dashboard-plugin",
     "@kbn/cloud",
     "@kbn/config",
-    "@kbn/core-http-server-mocks",
   ]
 }
